% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processNC-package.R
\docType{data}
\name{bavaria}
\alias{bavaria}
\title{Global Administrative Boundary of Bavaria}
\format{
\code{sp::SpatialPolygonsDataFrame}
}
\description{
Global Administrative Boundary (GADM) of Bavaria
}
\details{
This dataset is a shapefile containing a single polygon 
of the administrative boundary of Bavaria.
}
