% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGrid.R
\name{convertGrid}
\alias{convertGrid}
\title{Convert grid type of NetCDF file}
\usage{
convertGrid(basedir)
}
\arguments{
\item{basedir}{\code{character}. Path to a file directory containing one or multiple NetCDF files.}
}
\value{
The function overwrites the existing NetCDF files with a new file with the correct grid type, so no new file is returned.
}
\description{
Convert grid type of NetCDF file
}
\examples{
\dontrun{
extdir <- paste0(system.file(package="processNC"), "/extdata")
convertGrid(basedir=extdir)
}
}
