% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeNC.R
\name{mergeNC}
\alias{mergeNC}
\title{Merge multiple NetCDF files into one}
\usage{
mergeNC(files, outfile)
}
\arguments{
\item{files}{\code{character}. List of filepaths, which should lead to NetCDF files.}

\item{outfile}{\code{character}. Output filename of the merged data.}
}
\value{
A NetCDF file containing all of the merged data.
}
\description{
Merge multiple NetCDF files into one
}
