% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remapNC.R
\name{remapNC}
\alias{remapNC}
\title{Remap NetCDF file}
\usage{
remapNC(gridfile, infile, outfile)
}
\arguments{
\item{gridfile}{\code{character}. Filepath to the grid.txt file.}

\item{infile}{\code{character}. Filepath, which should lead to NetCDF files.}

\item{outfile}{\code{character}. Output filename of the remapped data.}
}
\value{
A NetCDF file containing the croped data.
}
\description{
Remap NetCDF file
}
\examples{
\dontrun{
file <- list.files(paste0(system.file(package="processNC"), "/extdata"), 
                   pattern="tas.*\\\\.nc", full.names=TRUE)[1]
temp <- tempfile(fileext=".nc")
gridfile <- paste0(system.file(package="processNC"), "/extdata/euro-cordex_grid_bav.txt")
remapNC(gridfile=gridfile, infile=file, outfile=temp)
raster::plot(raster::stack(temp)) 
}
}
