% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setgridtypeNC.R
\name{setgridtypeNC}
\alias{setgridtypeNC}
\title{Set grid type of NetCDF file}
\usage{
setgridtypeNC(type = "curvilinear", infile, outfile)
}
\arguments{
\item{type}{\code{character}. Grid type}

\item{infile}{\code{character}. Filepath, which should lead to NetCDF files.}

\item{outfile}{\code{character}. Output filename of the remapped data.}
}
\value{
A NetCDF file containing the croped data.
}
\description{
Set grid type of NetCDF file
}
\examples{
\dontrun{
file <- list.files(paste0(system.file(package="processNC"), "/extdata"), 
                   pattern="tas.*\\\\.nc", full.names=TRUE)[1]
temp <- tempfile(fileext=".nc")
setgridtypeNC(type="curvilinear", infile=file, outfile=temp)
raster::plot(raster::stack(temp)) 
}
}
