% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleCDO.R
\name{simpleCDO}
\alias{simpleCDO}
\title{Run simple CDO commands}
\usage{
simpleCDO(command, infile)
}
\arguments{
\item{command}{\code{character}. Filepath to the grid.txt file.}

\item{infile}{\code{character}. Filepath, which should lead to NetCDF files.}
}
\value{
Output
}
\description{
Run simple CDO commands
}
\examples{
\dontrun{
file <- list.files(paste0(system.file(package="processNC"), "/extdata"), 
                   pattern="tas.*\\\\.nc", full.names=TRUE)[1]
simpleCDO(command="showname", infile=file)
simpleCDO(command="showtimestamp", infile=file)
simpleCDO(command="sinfo", infile=file)
}
}
