% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetNC.R
\name{subsetNC}
\alias{subsetNC}
\title{Subset NetCDF files by x,y,time or variable}
\usage{
subsetNC(
  files,
  startdate = NA,
  enddate = NA,
  ext = "",
  varid = "",
  filename = "",
  format = "raster",
  overwrite = FALSE
)
}
\arguments{
\item{files}{\code{character}. File path.}

\item{startdate}{\code{integer}. Start year.}

\item{enddate}{\code{integer}. End year.}

\item{ext}{\code{integer}. Extent object.}

\item{varid}{\code{character}. Variable that should be extracted.}

\item{filename}{\code{character}. Output filename.}

\item{format}{\code{character}. Output file type. See \code{\link[raster]{writeRaster}} for different Options. The default format is "raster".}

\item{overwrite}{\code{logical}. Should file be overwritten or not.}
}
\value{
A \code{numeric} raster stack with the cropped NetCDF data.
}
\description{
Subset a NetCDF file by a spatial extent (x,y), time or variable and 
output it as a raster stack
}
\examples{
files <- list.files(paste0(system.file(package="processNC"), "/extdata"), 
                    pattern="tas.*\\\\.nc", full.names=TRUE)
subsetNC(files, ext=c(8.5, 14, 47, 51), startdate=1990, enddate=1999)
}
