% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseRaster.R
\name{summariseRaster}
\alias{summariseRaster}
\title{Aggregate values of multiple NetCDF Files}
\usage{
summariseRaster(path = getwd(), mean = TRUE, error = TRUE,
  filename1 = NA, filename2 = NA, overwrite = FALSE)
}
\arguments{
\item{path}{\code{character}. Path of file to use.}

\item{mean}{\code{logical}.}

\item{error}{\code{logical}.}

\item{filename1}{\code{character}. Filename for average monthly raster stack}

\item{filename2}{\code{character}. Filename for monthly coefficient of variance raster stack}

\item{overwrite}{\code{logical}. Should file be overwritten or not.}
}
\value{
A \code{numeric} raster stack with monthly layers of 
aggregated data over specificed time period.
}
\description{
Aggregate spatio-temporal measurements of one raster stack with daily layers 
for multiple years to a raster stack with monthly layers
}
\examples{
\dontrun{
summariseRaster()
}
}
