% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateNC.R
\name{aggregateNC}
\alias{aggregateNC}
\title{Aggregate one or multiple NetCDF Files over time and/or space}
\usage{
aggregateNC(infile, outfile, var, startyear, endyear)
}
\arguments{
\item{infile}{\code{character}. A filepath or list of filepaths. Filepath should lead to a NetCDF file.}

\item{outfile}{\code{character}. Output filename of the averaged data.}

\item{var}{\code{character}. Variable which is used, one of tasmax, tasmin and pr.}

\item{startdate}{\code{integer}. Start year.}

\item{enddate}{\code{integer}. End year.}
}
\value{
A NetCDF with temporal values averaged over the specified area.
}
\description{
Aggregate one or multiple NetCDF Files over time and/or space
}
