% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processNC-package.R
\docType{package}
\name{processNC-package}
\alias{processNC-package}
\alias{processNCpackage}
\title{R Package for processing large NetCDF files}
\description{
R Package for processing large NetCDF files
}
\details{
Analysis of large NetCDF files (Cropping, Summarising) for 
further use with R using the raster package.
}
\author{
RS-eco
}
\keyword{package}
